LayoutTests/http/tests/model/model-document-interactive.html

<!doctype html>
<html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>

promise_test(async test => {
    const frame = document.createElement("iframe");
    frame.src = "UnitBox.usdz";
    document.body.appendChild(frame);

    await new Promise(resolve => frame.addEventListener("load", resolve, { once: true }));

    const model = frame.contentDocument.querySelector("model");
    assert_true(model.interactive);
}, "<model> in standalone document is interactive.");

</script>
</body>
</html>
